refactor: improve re-authentication by separating it for a new context

+++ dkmovie/src/pages/account/my-account/security-tabs/2fa-card/totp/remove-totp.tsx
@@ -13,13 +13,14 @@ import {
   AlertDialogTrigger,
 } from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
-import { useSession } from "@/hooks/use-session";
+import { useReAuthenticate } from "@/hooks/use-reauthenticate";
 import { deleteTOTP } from "@/http/account/2fa";
 import { needReAuthentication } from "@/utils/auth-flows";
 
 export function RemoveTOTP() {
   const queryClient = useQueryClient();
-  const { initializeReAuthentication, isReAuthenticating } = useSession();
+  const { initializeReAuthentication, isReAuthenticating } =
+    useReAuthenticate();
 
   const { mutate: deleteTOTPMutation, isPending: isDeletingTOTP } = useMutation(
     {
@@ -32,7 +33,9 @@ export function RemoveTOTP() {
       },
       onError: (error) => {
         if (needReAuthentication(error)) {
-          initializeReAuthentication(deleteTOTPMutation);
+          initializeReAuthentication({
+            onReAuthenticated: deleteTOTPMutation,
+          });
           return;
         }
 
@@ -42,44 +45,40 @@ export function RemoveTOTP() {
     },
   );
 
+  if (isReAuthenticating) return null;
+
   return (
-    !isReAuthenticating && (
-      <AlertDialog>
-        <AlertDialogTrigger asChild>
-          <Button type="button" variant="outline" size="sm">
-            Remove
-          </Button>
-        </AlertDialogTrigger>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Remove TOTP</AlertDialogTitle>
-            <AlertDialogDescription>
-              Are you sure you want to remove your TOTP?
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel asChild>
-              <Button type="button" variant="outline">
-                Cancel
-              </Button>
-            </AlertDialogCancel>
-            <AlertDialogAction asChild>
-              <Button
-                type="button"
-                variant="destructive"
-                disabled={isDeletingTOTP}
-                onClick={() => deleteTOTPMutation()}
-              >
-                {isDeletingTOTP ? (
-                  <Loader className="animate-spin" />
-                ) : (
-                  "Remove"
-                )}
-              </Button>
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
-    )
+    <AlertDialog>
+      <AlertDialogTrigger asChild>
+        <Button type="button" variant="outline" size="sm">
+          Remove
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Remove TOTP</AlertDialogTitle>
+          <AlertDialogDescription>
+            Are you sure you want to remove your TOTP?
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel asChild>
+            <Button type="button" variant="outline">
+              Cancel
+            </Button>
+          </AlertDialogCancel>
+          <AlertDialogAction asChild>
+            <Button
+              type="button"
+              variant="destructive"
+              disabled={isDeletingTOTP}
+              onClick={() => deleteTOTPMutation()}
+            >
+              {isDeletingTOTP ? <Loader className="animate-spin" /> : "Remove"}
+            </Button>
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 }
